Add BundleGraphDiff component (#17)
This implements the `BundleGraphDiff` component and adds it to the
sample app. Given two `BundleStats` objects, it will produce a diff
graph which shows nodes and edges that were added or removed.

Because chunk group IDs are not necessarily stable between builds, this
uses the chunk group name to match chunk groups between the input
graphs. Unnamed chunk groups are ignored. Theoretically we should be
able to use a chunk group's `origins` information to match unnamed
chunks, but that will be a future improvement.